bug 1269104: Move kuma.wiki fixtures to conftest
